Some significant perormance improvements
By moving the gE and gI matrices to static storage, we can convert a
call to malloc to a call to memset, which is ridiculously faster.

Also, removing some of the annoying printf statements has a significant
improvement.

Current speed is 6s to run 1s of simulation of a 100x100 grid