Fix up some float/double conversions
Since we are now using the magic Compute_float typedef - it makes sense
to try and fix up some automaticd conversions that the compiler is
doing.  Most of these are due to the fact that any literal constant is a
double.  I can see no easy way to fix this so I haven't made changes in
Parameters.h and am probably going to run with the warning off, but it
would be a good idea to frequently check to see that we aren't missing
any errors