Allow gE/gI to be viewed in matlab
Previous changes had made gE and gI invisible to the matlab code as it
was hidden inside a static variable in the function.

Change the code so that we now create a

    const volatile compute_float* const

to the data which is read only.  (I should possibly be using C++ for
this, but whatever works is best).