Remove requirement to #include parameters.h
Previously, parameters.h was getting included everywhere.  This was
really bad, we really should have been including paramheader.h instead.
This is now fixed so that we include paramheader (which is basically
just typedefs) instead.  Unfortunately, the sizes of some arrays need to be compile
time constants (#define or enums) and const int is not good enough, so
there is a pretty nasty hack to get the sizes working properly.